add in some corrections
